backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
sentence-transformers>=2.7.0
chromadb==0.4.18
openai==1.3.7
python-dotenv==1.0.0
pypdf==4.0.0
beautifulsoup4==4.12.2
requests==2.31.0
numpy>=1.24.0
pandas>=2.2.0
tiktoken>=0.6.0
markdown==3.5.1
setuptools>=65.0.0
